+++ examples/i2s_slave/src/i2s_slave_sw_pll.c
@@ -178,6 +178,7 @@ void sw_pll_test(void){
     sw_pll_init(&sw_pll,
                 SW_PLL_15Q16(0.0),
                 SW_PLL_15Q16(1.0),
+                SW_PLL_15Q16(0.0),
                 CONTROL_LOOP_COUNT,
                 PLL_RATIO,
                 BCLKS_PER_LRCLK,
@@ -189,7 +190,7 @@ void sw_pll_test(void){
                 PPM_RANGE);
 
 
-    printf("i_windup_limit: %ld\n", sw_pll.i_windup_limit);
+    printf("i_windup_limit: %ld\n", sw_pll.pi_state.i_windup_limit);
 
 
     // Initialise app_data